requirements.txt
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9